requirements.txt
numpy
matplotlib
scikit-image
progressbar2
pytest
pytest-runner
coverage
sphinx
recommonmark
sphinx_rtd_theme
sphinx-autodoc-typehints